Les Dao + les cruds

+++ src/GoldenCage/Dao/AdminDAO.java
@@ -8,13 +8,10 @@ package GoldenCage.Dao;
 
 import GoldenCage.entites.Comptes;
 import GoldenCage.entites.Admin;
-import GoldenCage.util.Connexion;
-import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
-import java.util.List;
+import java.util.List; 
 
 /**
  *
@@ -25,49 +22,20 @@ public class AdminDAO extends ComptesDAO{
     public void insertAdmin(Comptes c,Admin a){
         
         super.insertCompte(c);        
-        String requete = "insert into admin values (?,?)";
-        try {
-            PreparedStatement ps = Connexion.getInstance().prepareStatement(requete);
-            ps.setInt(1, a.getIdAdmin());
-            ps.setInt(2, a.getIdCompte());
-            ps.executeUpdate();
-            System.out.println("Ajout effectuée avec succès");
-        } catch (SQLException ex) {
-           //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
-            System.out.println("Erreur lors de l'insertion "+ex.getMessage());
-        }
+        String requete = "insert into admin values ("+a.getIdAdmin()+","+a.getIdCompte()+")";
+        cr.insert(requete);
     }
     
     public void updateadmin(Admin a){
-        String requete = "update comptes set nom=?,prenom=?,adresse=?,email=?,numTel=? where idCompte=?";
-        try {
-            PreparedStatement ps = Connexion.getInstance().prepareStatement(requete);
-            ps.setString(1, a.getNom());
-            ps.setString(2, a.getPrenom());
-            ps.setString(3, a.getAdresse());
-            ps.setString(4, a.getEmail());
-            ps.setInt(5, a.getNumTel());
-            ps.setInt(6, a.getIdCompte());
-            ps.executeUpdate();
-            System.out.println("Mise à jour effectuée avec succès");
-        } catch (SQLException ex) {
-           //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
-            System.out.println("Erreur lors de la mise à jour "+ex.getMessage());
-        }
+        String requete = "update comptes set nom='"+a.getNom()+"',prenom='"+a.getPrenom()+"',adresse='"
+                +a.getAdresse()+"',email='"+a.getEmail()+"',numTel="+a.getNumTel()+",pseudo='"
+                +a.getPseudo()+"',motDePass='"+a.getMotDePasse()+"' where idCompte="+a.getIdCompte();
+        cr.update(requete);
     }
     
      public void deleteAdmin(Admin a){
-        String requete = "delete from admin where idAdmin=?";
-        try {
-            PreparedStatement ps = Connexion.getInstance().prepareStatement(requete);
-            ps.setInt(1, a.getIdAdmin());
-            ps.executeUpdate();
-            super.deleteCompte(a.getIdCompte());
-            System.out.println("Admin supprimée");
-        } catch (SQLException ex) {
-           //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
-            System.out.println("Erreur lors de la suppression "+ex.getMessage());
-        }
+        String requete = "delete from admin where idAdmin="+a.getIdAdmin();
+        cr.delete(requete);
     }
      
      public List<Comptes> afficherAdmin (){
@@ -76,8 +44,8 @@ public class AdminDAO extends ComptesDAO{
 
         String requete = "select * from Comptes where typeCompte='ADMIN'";
         try {
-           Statement statement = Connexion.getInstance().createStatement();
-            ResultSet resultat = statement.executeQuery(requete);
+           
+            ResultSet resultat = cr.list(requete);
 
             while(resultat.next()){
                 Comptes comp =new Comptes();
@@ -92,7 +60,6 @@ public class AdminDAO extends ComptesDAO{
             }
             return compteadmin;
         } catch (SQLException ex) {
-           //Logger.getLogger(PersonneDao.class.getName()).log(Level.SEVERE, null, ex);
             System.out.println("Erreur lors du chargement des comptes des admin "+ex.getMessage());
             return null;
         }
